requirements.txt
flask-appconfig>=0.10
flask-bootstrap
flask-nav
flask-debug
flask-wtf
flask_table